requirements.txt
attrs==23.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
confection==0.1.5
cymem==2.0.8
Cython==0.29.37
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
medspacy==1.2.0
medspacy_quickumls==3.2
medspacy_unqlite==0.9.6
murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pathlib_abc==0.1.1
pathy==0.11.0
pluggy==1.5.0
preshed==3.0.9
psycopg2-binary==2.9.9
pydantic==1.10.15
PyFastNER==1.0.10
PyRuSH==1.0.8
pysbd==0.3.4
pysimstring==1.2.1
pytest==8.2.1
python-dateutil==2.9.0.post0
pytz==2024.1
quicksectx==0.3.9
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
smart-open==6.4.0
spacy==3.5.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.1.12
tomli==2.0.1
tqdm==4.66.4
typer==0.9.4
typing_extensions==4.12.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.3
